Update identifier URL and links section with all references
The identifier (aka CVE ID) URL was previously being set to the first link in the references array. However since it is specifically the CVE, it should always be set to the MITRE CVE URL.

Additionally, the links array only had the MITRE CVE URL. Now, the links array has all the references available from the original JSON references array.