Update scaGitLabScript.py
There is discrepancy between the severity reported via semgrep tool (after running semgrep command to fetch json output for sca) and the output after running scaGitLabScript.py (Convert the json into Gitlab digestable format). The discrepancy in severity is due to how the severity of reachable vulnerabilities is handled. Based on the logic below, if kind == reachable, the severity is set to Critical. Suggesting changes.